Separate Output ID Proof and Merkle Tree schemas
